requirements.txt
openai
python-dotenv
wcmatch
tiktoken
scikit-learn
pytest
pydantic
numpy
chardet
pathspec
